requirements.txt
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
fastapi>=0.110.0
uvicorn>=0.27.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0
colorama>=0.4.6
boto3>=1.34.0
sqlalchemy>=2.0.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
cryptography>=42.0.0
